:recycle: [refactor] 비디오 디테일 screen 리팩토링 (HH-453)
