+++ lib/ui/view/profile/profile_tab_videos_view.dart
@@ -3,7 +3,7 @@ import 'package:pocket_pose/data/entity/request/profile_videos_request.dart';
 import 'package:pocket_pose/data/entity/response/profile_response.dart';
 import 'package:pocket_pose/data/local/provider/multi_video_play_provider.dart';
 import 'package:pocket_pose/data/remote/provider/profile_provider.dart';
-import 'package:pocket_pose/ui/screen/profile/profile_video_detail_screen.dart';
+import 'package:pocket_pose/ui/screen/video/video_detail_screen.dart';
 import 'package:pocket_pose/ui/loader/profile_video_skeleton_loader.dart';
 import 'package:pocket_pose/ui/widget/page_route_with_animation.dart';
 import 'package:provider/provider.dart';
